Refactor so SSL parts are in a separate file
